// AngularMain/src/app/modules/models/OffersDto.ts
export interface OffersDto
{
    tenentId: number;
        serviceId: number;
        active: string;
        crupId: number;
        userid: string;
        entrydate: Date;
        entrytime: Date;
        updttime: Date;
        uploadDate: Date;
        uploadby: string;
        syncDate: Date;
        syncby: string;
        synId: number;
        file1: string;
        offerImage: string;
        offerType: string;
        offer: string;
        offerStartDate: Date;
        offerEndDate: Date;
        offerAmount: number;
        offerTypeName:string;
        electronicForm1File: string,
  electronicForm2File:string,
  offerImageFile: string
}